refactor: move utilities to `src/util.ts`
